implement proto_output_to_native
